--- tests/test_version.py
@@ -1,11 +1,0 @@
-from click.testing import CliRunner
-
-from mmgmt.cli import cli
-
-
-def test_version():
-    runner = CliRunner()
-    with runner.isolated_filesystem():
-        result = runner.invoke(cli, ["--version"])
-        assert result.exit_code == 0
-        assert result.output.startswith("cli, version ")
